+++ include/config.h
@@ -244,6 +244,7 @@ enum ProjectIdentifier {
     regex_music,
     regex_encounter_types,
     regex_terrain_types,
+    regex_incbin,
     pals_output_extension,
     tiles_output_extension,
 };