+++ src/config.cpp
@@ -5,6 +5,7 @@
 #include "validator.h"
 #include "utility.h"
 #include "metatile.h"
+#include "regex.h"
 #include <QDir>
 #include <QFile>
 #include <QFormLayout>
@@ -133,6 +134,7 @@ const QMap<ProjectIdentifier, QPair<QString, QString>> ProjectConfig::defaultIde
     {ProjectIdentifier::regex_music,                   {"regex_music",                   "\\b(SE|MUS)_"}},
     {ProjectIdentifier::regex_encounter_types,         {"regex_encounter_types",         "\\bTILE_ENCOUNTER_"}},
     {ProjectIdentifier::regex_terrain_types,           {"regex_terrain_types",           "\\bTILE_TERRAIN_"}},
+    {ProjectIdentifier::regex_incbin,                  {"regex_incbin",                  Regex::Pattern_INCBIN}},
     // Other
     {ProjectIdentifier::pals_output_extension,         {"pals_output_extension",         ".gbapal"}},
     {ProjectIdentifier::tiles_output_extension,        {"tiles_output_extension",        ".4bpp.lz"}},